Update the grammar for raw strings
There are two different grammars for raw strings in the specification, one that allows lone backslashes and one that does not allow backslash unless followed by single quote.

Not allowing backslashes unless followed by single quote is contradicted by the examples in the specification and by the compliance tests (specifically '\u03a6' is given as a valid raw string in both).

Allowing lone backslashes is problematic since it makes the semantics of '\\' unclear. In the current go implementation, for example, this is treated as an unterminated raw string, but it could equally well be interpreted as a string with two backslashes by another implementation, and neither would be wrong.

This clarifies that we want to treat backslashes literally, except when followed by single quote. Also, to make it possible to end a string with a backslash we also require backslashes to be escaped.